refactor code for sort class path
